fix : shareGroupId : Int -> Long 수정 GroupDetailFolderScreen 수정중

+++ app/src/main/java/com/hgh/na_o_man/data/dto/share_group/response/ShareGroupInfoDto.kt
@@ -8,7 +8,7 @@ data class ShareGroupInfoDto(
     val inviteUrl: String,
     val memberCount: Int,
     val name: String,
-    val shareGroupId: Int
+    val shareGroupId: Long
 ) {
     fun toModel() = ShareGroupInfoModel(
         createdAt = this.createdAt,
